Split `sshmux.go`
